Fix and normalize handling of none matching highlight fields (BC Break) (#525)
There is inconsistent between how the result is if there are multiple
"highlight" fields and the a highlight field does not have a match.

- Elasticsearch, Opensearch, Solr: Do return nothing if the field has no
match and so no highlighting is applied
 - Meilisearch, Algolia, Loupe: Returns the fields value
 
To be consistent I decided to go with the Elasticsearch behaviour
because it seems my point of view easier for enduser to write:
`$document['_formatted']['description'] ?? $document['description']`.
As the other way `\str_contains($document['_formatted']['description'],
$search->highlightPreTag) ? $document['_formatted']['description'] :
null`.

Another issue appeared which is #524 that none existing fields did throw
an error on elastic and opensearch, we can not be as structure strict as
I thought we can here.

fixes #524