Add possibility that an Index can have options (#541)
Maybe there are adapter specific options for an index. Like we have
already `options` array for `fields` we add here it for Index. So the
`SchemaManager` of adapters like `Loupe` can example configure a
`TypoTolerance` or other options.

Target of this PR is only add possibility to have options not yet any
adapter specific implementation for it. It can be used for projects to add additonal infos like a label  like in #539.

fixes part of #540