Add support for JsonObjectField (#616)
This adds a new field type which is just here to store metadata in a
json object (`array<string, mixed>`). This kind of data is not search,
facet or filterable.

Usage:

```php
new Index('test', [
    'uuid' => new Field\IdentifierField('uuid'),
    'title' => new Field\TextField('title', sortable: true),
    'metadata' => new Field\JsonObjectField('metadata'),
]);
```

```php
$engine->saveDocument([
    'uuid' => $uuid,
    'title' => $title,
    'metadata' => [
        'routeAttributes' => [
            'site' => 'example',
        ],
    ],
]);
```

See https://github.com/PHP-CMSIG/search/issues/614

This feature is experimental and may change in future releases.