Mention rollerworks in alternative or similar projects (#628)
Why it is not directly focused on search engine or the index, reindex
part of it or how even the data is put into the source. I think it still
worth mentioning the project by @sstok also in the similar alternatives:

https://github.com/rollerworks/search

What could maybe be possible that somebody create a Rollerworks
Extension supporting also [SEALs Search &
Filters](https://php-cmsig.github.io/search/search-and-filters/index.html).

__Why even care about other projects?__

The idea did came up why first time I did read the README of
[thephpleague/object-mapper](https://github.com/thephpleague/object-mapper?tab=readme-ov-file#alternatives).
Who listed alternatives and think that is how Open Source should be,
about sharing, cooperations and friendly coexisting. There exists always
multiple solutions for the same problem and if there are such solutions
we know we should share them with others.